// src/service/xrpl-meta/enum/tokens-sort.enum.ts
export enum ETokensSort {
  trustlines_delta_24h = 'trustlines_delta_24h',
  trustlines_percent_24h = 'trustlines_percent_24h',
  trustlines_delta_7d = 'trustlines_delta_7d',
  trustlines_percent_7d = 'trustlines_percent_7d',
  holders = 'holders',
  holders_delta_24h = 'holders_delta_24h',
  holders_percent_24h = 'holders_percent_24h',
  holders_delta_7d = 'holders_delta_7d',
  holders_percent_7d = 'holders_percent_7d',
  supply = 'supply',
  supply_delta_24h = 'supply_delta_24h',
  supply_percent_24h = 'supply_percent_24h',
  supply_delta_7d = 'supply_delta_7d',
  supply_percent_7d = 'supply_percent_7d',
  marketcap = 'marketcap',
  marketcap_delta_24h = 'marketcap_delta_24h',
  marketcap_percent_24h = 'marketcap_percent_24h',
  marketcap_delta_7d = 'marketcap_delta_7d',
  marketcap_percent_7d = 'marketcap_percent_7d',
  price_percent_24h = 'price_percent_24h',
  price_percent_7d = 'price_percent_7d',
  volume_24h = 'volume_24h',
  volume_7d = 'volume_7d',
  exchanges_24h = 'exchanges_24h',
  exchanges_7d = 'exchanges_7d',
  takers_24h = 'takers_24h',
  takers_7d = 'takers_7d',
  trustlines = 'trustlines',
}